Cleaned some stuff up
